TASK: Update some unit tests (#1869)
* TASK: Update unit tests
